organizando o codigo
